changed default backend from biber to bibtex
Biber by default, on a macOS environment running vscode and MacTex,  fails latex build of this project by not recognizing any reference or citation. And from the error logs it's next to impossible to find out why the project doesn't build. 

Not to mention bad user experience if the first time for newbie users building the project fails.